+++ src/main/java/com/likelion/allForOne/domain/tblCode/dto/CodeDto.java
@@ -1,4 +1,4 @@
-package com.likelion.allForOne.tblCode.dto;
+package com.likelion.allForOne.domain.tblCode.dto;
 
 import lombok.Builder;
 import lombok.Getter;
