+++ src/main/java/com/likelion/allForOne/domain/tblGroup/dto/GroupDto.java
@@ -1,4 +1,4 @@
-package com.likelion.allForOne.tblGroup.dto;
+package com.likelion.allForOne.domain.tblGroup.dto;
 
 import lombok.Builder;
 import lombok.Getter;
